migrating tailwindcss-colourstyling to shared endpoint
