+++ pages/api/studentPortal/lessons/web/react/index.ts
@@ -1,6 +1,7 @@
 import { getDataForComponentsLesson } from "./components";
 import { getGithubLessonData } from "./github";
 import { getPropsLessonData } from "./props";
+import { getTailwindColourStylingLessonData } from "./tailwindcss-colourstyling";
 
 export function getLessonForReactCourse(lessonId: string) {
   if (lessonId === "components") {
@@ -9,6 +10,8 @@ export function getLessonForReactCourse(lessonId: string) {
     return getGithubLessonData();
   } else if (lessonId === "props") {
     return getPropsLessonData();
+  }else if (lessonId === "tailwindcss-colourstyling") {
+    return getTailwindColourStylingLessonData();
   }
 
   throw new Error(`Could not find lessons`);
